requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
Django==5.1.7
djangorestframework==3.15.2
favicon==0.7.0
humanize==4.12.1
idna==3.10
lxml==5.3.1
packaging==24.2
pikepdf==9.5.2
pillow==11.1.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
urltitle==0.3.8
wrapt==1.17.2
